--- a/R/BoxCox.R
+++ b/R/BoxCox.R
@@ -131,13 +131,13 @@ prep.step_BoxCox <- function(x, training, info = NULL, ...) {
 
 #' @export
 bake.step_BoxCox <- function(object, new_data, ...) {
-  param <- names(object$lambdas)
-  check_new_data(param, object, new_data)
+  col_names <- names(object$lambdas)
+  check_new_data(col_names, object, new_data)
 
-  for (i in seq_along(object$lambdas)) {
-    new_data[[param[i]]] <- bc_trans(
-      new_data[[param[i]]],
-      lambda = object$lambdas[i]
+  for (col_name in col_names) {
+    new_data[[col_name]] <- bc_trans(
+      new_data[[col_name]],
+      lambda = object$lambdas[col_name]
     )
   }