--- a/R/date.R
+++ b/R/date.R
@@ -241,12 +241,14 @@ get_date_features <-
 
 #' @export
 bake.step_date <- function(object, new_data, ...) {
-  check_new_data(names(object$columns), object, new_data)
+  col_names <- names(object$columns)
+  check_new_data(col_names, object, new_data)
 
   new_cols <- rep(
     length(object$features),
-    each = length(object$columns)
+    each = length(col_names)
   )
+  names(new_cols) <- col_names
 
   date_values <- matrix(NA, nrow = nrow(new_data), ncol = sum(new_cols))
 
@@ -258,11 +260,11 @@ bake.step_date <- function(object, new_data, ...) {
   new_names <- vector("character", length = ncol(date_values))
 
   strt <- 1
-  for (i in seq_along(object$columns)) {
-    cols <- (strt):(strt + new_cols[i] - 1)
+  for (col_name in col_names) {
+    cols <- (strt):(strt + new_cols[col_name] - 1)
 
     tmp <- get_date_features(
-      dt = new_data[[object$columns[i]]],
+      dt = new_data[[col_name]],
       feats = object$features,
       locale = object$locale %||% Sys.getlocale("LC_TIME"),
       abbr = object$abbr,
@@ -272,11 +274,7 @@ bake.step_date <- function(object, new_data, ...) {
 
     date_values[, cols] <- tmp
 
-    new_names[cols] <- paste(
-      object$columns[i],
-      names(tmp),
-      sep = "_"
-    )
+    new_names[cols] <- paste(col_name, names(tmp), sep = "_")
 
     strt <- max(cols) + 1
   }
@@ -287,7 +285,7 @@ bake.step_date <- function(object, new_data, ...) {
 
   new_data <- vec_cbind(new_data, date_values)
 
-  new_data <- remove_original_cols(new_data, object, object$columns)
+  new_data <- remove_original_cols(new_data, object, col_names)
   new_data
 }