--- a/R/factor2string.R
+++ b/R/factor2string.R
@@ -90,9 +90,12 @@ prep.step_factor2string <- function(x, training, info = NULL, ...) {
 
 #' @export
 bake.step_factor2string <- function(object, new_data, ...) {
-  check_new_data(names(object$columns), object, new_data)
+  col_names <- names(object$columns)
+  check_new_data(col_names, object, new_data)
 
-  new_data[, object$columns] <- map(new_data[, object$columns], as.character)
+  for (col_name in col_names) {
+    new_data[[col_name]] <- as.character(new_data[[col_name]])
+  }
 
   new_data
 }