--- a/R/geodist.R
+++ b/R/geodist.R
@@ -208,9 +208,10 @@ geo_dist_calc_lat_lon <- function(x_1, y_1, x_2, y_2, earth_radius = 6371e3,
 
 #' @export
 bake.step_geodist <- function(object, new_data, ...) {
-  check_new_data(names(object$columns), object, new_data)
+  col_names <- names(object$columns)
+  check_new_data(col_names, object, new_data)
 
-  if (length(object$columns) == 0) {
+  if (length(col_names) == 0) {
     return(new_data)
   }
 
@@ -219,16 +220,16 @@ bake.step_geodist <- function(object, new_data, ...) {
   if (object$is_lat_lon) {
     dist_vals <-
       geo_dist_calc_lat_lon(
-        new_data[[object$columns[2]]], # lon
-        new_data[[object$columns[1]]], # lat
+        new_data[[col_names[2]]], # lon
+        new_data[[col_names[1]]], # lat
         object$ref_lon,
         object$ref_lat
       )
   } else {
     dist_vals <-
       geo_dist_calc_xy(
-        new_data[[object$columns[2]]], # lon
-        new_data[[object$columns[1]]], # lat
+        new_data[[col_names[2]]], # lon
+        new_data[[col_names[1]]], # lat
         object$ref_lon,
         object$ref_lat
       )