--- a/R/impute_lower.R
+++ b/R/impute_lower.R
@@ -150,17 +150,21 @@ prep.step_lowerimpute <- prep.step_impute_lower
 
 #' @export
 bake.step_impute_lower <- function(object, new_data, ...) {
-  check_new_data(names(object$threshold), object, new_data)
+  col_names <- names(object$threshold)
+  check_new_data(col_names, object, new_data)
 
-  for (i in names(object$threshold)) {
-    affected <- which(new_data[[i]] <= object$threshold[[i]])
+  for (col_name in col_names) {
+    threshold <- object$threshold[[col_name]]
+    affected <- which(new_data[[col_name]] <= threshold)
+
     if (length(affected) > 0) {
-      new_data[[i]][affected] <- runif(
+      new_data[[col_name]][affected] <- runif(
         length(affected),
-        max = object$threshold[[i]]
+        max = threshold
       )
     }
   }
+
   new_data
 }