--- a/R/impute_median.R
+++ b/R/impute_median.R
@@ -146,14 +146,17 @@ prep.step_medianimpute <- prep.step_impute_median
 
 #' @export
 bake.step_impute_median <- function(object, new_data, ...) {
-  check_new_data(names(object$medians), object, new_data)
+  col_names <- names(object$medians)
+  check_new_data(col_names, object, new_data)
 
-  for (i in names(object$medians)) {
-    if (any(is.na(new_data[[i]]))) {
-      new_data[[i]] <- vec_cast(new_data[[i]], object$medians[[i]])
+  for (col_name in col_names) {
+    median <- object$medians[[col_name]]
+    if (any(is.na(new_data[[col_name]]))) {
+      new_data[[col_name]] <- vec_cast(new_data[[col_name]], median)
     }
-    new_data[is.na(new_data[[i]]), i] <- object$medians[[i]]
+    new_data[is.na(new_data[[col_name]]), col_name] <- median
   }
+
   new_data
 }