--- a/R/isomap.R
+++ b/R/isomap.R
@@ -192,9 +192,10 @@ prep.step_isomap <- function(x, training, info = NULL, ...) {
 
 #' @export
 bake.step_isomap <- function(object, new_data, ...) {
-  check_new_data(names(object$columns), object, new_data)
+  col_names <- names(object$columns)
+  check_new_data(col_names, object, new_data)
 
-  if (object$num_terms > 0 && length(object$columns) > 0L) {
+  if (object$num_terms > 0 && length(col_names) > 0L) {
     isomap_vars <- colnames(environment(object$res@apply)$indata)
     suppressMessages({
       comps <- object$res@apply(