--- a/R/isomap.R
+++ b/R/isomap.R
@@ -196,18 +196,18 @@ bake.step_isomap <- function(object, new_data, ...) {
   check_new_data(col_names, object, new_data)
 
   if (object$num_terms > 0 && length(col_names) > 0L) {
-    isomap_vars <- colnames(environment(object$res@apply)$indata)
     suppressMessages({
       comps <- object$res@apply(
-        dimred_data(new_data[, isomap_vars, drop = FALSE])
+        dimred_data(new_data[, col_names, drop = FALSE])
       )@data
     })
     comps <- comps[, seq_len(object$num_terms), drop = FALSE]
     comps <- as_tibble(comps)
     comps <- check_name(comps, new_data, object)
     new_data <- vec_cbind(new_data, comps)
-    new_data <- remove_original_cols(new_data, object, isomap_vars)
+    new_data <- remove_original_cols(new_data, object, col_names)
   }
+
   new_data
 }