--- a/R/logit.R
+++ b/R/logit.R
@@ -99,13 +99,13 @@ pre_logit <- function(x, eps = 0) {
 
 #' @export
 bake.step_logit <- function(object, new_data, ...) {
-  check_new_data(names(object$columns), object, new_data)
+  col_names <- names(object$columns)
+  check_new_data(col_names, object, new_data)
 
-  for (i in seq_along(object$columns)) {
-    new_data[[object$columns[i]]] <-
-      binomial()$linkfun(
-        pre_logit(new_data[[object$columns[i]]], object$offset)
-      )
+  for (col_name in col_names) {
+    new_data[[col_name]] <- binomial()$linkfun(
+      pre_logit(new_data[[col_name]], object$offset)
+    )
   }
   new_data
 }