--- a/R/missing.R
+++ b/R/missing.R
@@ -113,7 +113,6 @@ prep.check_missing <- function(x, training, info = NULL, ...) {
 #' @export
 bake.check_missing <- function(object, new_data, ...) {
   col_names <- object$columns
-
   check_new_data(col_names, object, new_data)
 
   subset_to_check <- new_data[col_names]