… R/naomit.R
… R/newvalues.R
… R/normalize.R
… R/novel.R
… R/num2factor.R
… R/ordinalscore.R
… R/other.R
… R/percentile.R
… R/pls.R
… R/poly.R
… R/poly_bernstein.R
… R/range.R
… R/range_check.R
… R/relevel.R
… R/relu.R
… R/scale.R
… R/shuffle.R
… R/spatialsign.R
… R/spline_b.R
… R/spline_convex.R
… R/spline_monotone.R
… R/spline_natural.R
… R/spline_nonnegative.R
… R/sqrt.R
… R/string2factor.R
… R/time.R
… R/unknown.R
… R/unorder.R