This is quite a big PR.

it tries to give a more strieght forward unified approach to how to do things in `bake_step_*()` functions.

The gist of this, is that for many cases, we want to have the functions use the following structure

```r
for (col_name in col_names) {
  do something with  new_data[[col_name]]
}
```

It also has the benefit that the column names `col_names` are only extracted once, so we will gain a little speed and consistency that way.

Getting rid of `map()` inside `bake_step_*()` functions is also beneficial since they mess up the error handling/passing we are trying to use.

# Todo

The following steps needs these changes but for various reason didn't get done now. I'm planning on addressing these in separate PRs first.

- [x] `step_window()` https://github.com/tidymodels/recipes/pull/1172
    - is a 1 to many step, but doens’t act like it
- [x] `step_regex()` https://github.com/tidymodels/recipes/pull/1169
    - similar to step_dummy(), missing keep_orginal_cols()
- [x] `step_ratio()` https://github.com/tidymodels/recipes/pull/1173
    - doesn’t feel right
- [x] `step_profile()` https://github.com/tidymodels/recipes/pull/1177
    - also feels weird
- [x] `step_ns()` https://github.com/tidymodels/recipes/pull/1178
    - more work
- [x] `step_lag()` https://github.com/tidymodels/recipes/pull/1179
    - is a 1 to many step, but doesn’t act like it
- [x] `step_impute_linear()`, `step_impute_knn()`, `step_impute_bag()` https://github.com/tidymodels/recipes/pull/1184
    - missing lots of things. such as recipes_eval_select()
- [x] `step_depth()` https://github.com/tidymodels/recipes/pull/1180
    - missing check_new_data()
- [x] `step_count() ` https://github.com/tidymodels/recipes/pull/1169
    - is similar to step_regex()
- [x] `step_classdist()` https://github.com/tidymodels/recipes/pull/1183
    - similar to step_dummy() missing keep_orginal_cols()
- [x] `step_bs()` https://github.com/tidymodels/recipes/pull/1178
    - more work
- [x] `step_dummy()` https://github.com/tidymodels/recipes/pull/1182

## example benchmark

```r
old <- function(new_data, object) {
  if (length(object$columns) > 0) {
    for (i in seq_along(object$columns)) {
      new_data[, object$columns[i]] <-
        sample(getElement(new_data, object$columns[i]))
    }
  }
  new_data
}

new <- function(new_data, object) {
  col_names <- object$columns
  for (col_name in col_names) {
    new_data[[col_name]] <- sample(new_data[[col_name]])
  }
  new_data
}

object <- list(columns = colnames(mtcars))

bench::mark(
  check = FALSE,
  old(mtcars, object),
  new(mtcars, object)
)
#> # A tibble: 2 × 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old(mtcars, object)  158.5µs    167µs     5772.    4.34MB     41.4
#> 2 new(mtcars, object)   97.2µs    104µs     9231.  169.21KB     41.9
```